+++ yaks/lib/yaks/reader/json_api.rb
@@ -4,8 +4,8 @@ module Yaks
       def call(parsed_json, env = {})
         attributes = parsed_json['data'].first.dup
         links = attributes.delete('links') || {}
-        linked = parsed_json['linked'].nil? ? {} : parsed_json['linked'].dup
-        embedded   = convert_embedded(links, linked)
+        included = parsed_json['included'].nil? ? {} : parsed_json['included'].dup
+        embedded   = convert_embedded(links, included)
         Resource.new(
             type: Util.singularize(attributes.delete('type')[/\w+$/]),
             attributes: Util.symbolize_keys(attributes),
@@ -14,36 +14,34 @@ module Yaks
         )
       end
 
-      def convert_embedded(links, linked)
+      def convert_embedded(links, included)
         links.flat_map do |rel, link_data|
-          # If this is a compound link, the link_data will contain either
-          # * 'type' and 'id' for a one to one
-          # * 'type' and 'ids' for a homogeneous to-many relationship
-          # * 'data' being an array where each member has 'type' and 'id' for heterogeneous
-          if !link_data['type'].nil? && !link_data['id'].nil?
-            resource = linked.find{ |item| (item['id'] == link_data['id']) && (item['type'] == link_data['type']) }
-            call({'data'  => [resource], 'linked' => linked}).with(rels: [rel])
-          elsif !link_data['type'].nil? && !link_data['ids'].nil?
-            resources = linked.select{ |item| (link_data['ids'].include? item['id']) && (item['type'] == link_data['type']) }
-            members = resources.map { |r|
-              call({'data'  => [r], 'linked' => linked})
-            }
+          # A Link doesn't have to contain a `linkage` member.
+          # It can contain URLs instead, or as well, but we are only worried about *embedded* links here.
+          linkage = link_data['linkage']
+          # Resource linkage MUST be represented as one of the following:
+          #
+          # * `null` for empty to-one relationships.
+          # * a "linkage object" for non-empty to-one relationships.
+          # * an empty array ([]) for empty to-many relationships.
+          # * an array of linkage objects for non-empty to-many relationships.
+          if linkage.nil?
+            nil
+          elsif linkage.is_a? Array
             CollectionResource.new(
-                members: members,
-                type: link_data['type'],
-                rels: [rel]
-            )
-          elsif link_data['data'].present?
-            CollectionResource.new(
-                members: link_data['data'].map { |link|
-                  resource = linked.find{ |item| (item['id'] == link['id']) && (item['type'] == link['type']) }
-                  call({'data'  => [resource], 'linked' => linked})
+                members: linkage.map { |link|
+                  data = included.find{ |item| (item['id'] == link['id']) && (item['type'] == link['type']) }
+                  call({'data'  => [data], 'included' => included})
                 },
                 rels: [rel]
             )
+          else
+            data = included.find{ |item| (item['id'] == linkage['id']) && (item['type'] == linkage['type']) }
+            call({'data'  => [data], 'included' => included}).with(rels: [rel])
           end
         end.compact
       end
+
     end
   end
 end